feat: new structure for proker
