+++ app/pages/dashboard/(Kegiatan Muda-Mudi)/kegiatan-proker.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+  import ProkerBidangContainer from "~/features/proker/ProkerBidangContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("Kegiatan Muda-mudi / Proker");
+</script>
+
+<template>
+  <Title>Kegiatan Muda-mudi | Proker</Title>
+  <main>
+    <UCard>
+      <ProkerBidangContainer bidang="kegiatan_muda_mudi" />
+    </UCard>
+  </main>
+</template>
