+++ app/pages/dashboard/(Kurikulum)/kurikulum-monitoring-kehadiran-kelompok/index.vue
@@ -5,12 +5,12 @@
   import { kelasGenerusEnum } from "~~/shared/enum";
 
   const constantStore = useConstantStore();
-  constantStore.setTitle("Kurikulum / Monitoring Kehadiran Desa");
+  constantStore.setTitle("Kurikulum / Monitoring Kehadiran Kelompok");
   const authStore = useAuthStore();
 
   const namaKelas = ref<(typeof kelasGenerusEnum)[number]>("Muda-mudi");
   const { data: summary } = await useFetch(
-    `${APIBASE}/absensi-desa/daerah/summary`,
+    `${APIBASE}/absensi-generus/daerah/summary`,
     {
       query: {
         kelasPengajian: namaKelas,
@@ -28,7 +28,7 @@
   const searchDebounced = useDebounceFn((v) => {
     query.search = v;
   }, 300);
-  const { data, status } = await useFetch(`${APIBASE}/absensi-desa/daerah`, {
+  const { data, status } = await useFetch(`${APIBASE}/absensi-generus/daerah`, {
     query,
   });
