+++ app/pages/dashboard/(Sekretariat)/sekretariat-proker.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+  import ProkerBidangContainer from "~/features/proker/ProkerBidangContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("Sekretariat / Proker");
+</script>
+
+<template>
+  <Title>Sekretariat | Proker</Title>
+  <main>
+    <UCard>
+      <ProkerBidangContainer bidang="sekretariat" />
+    </UCard>
+  </main>
+</template>
