+++ app/pages/dashboard/(Tahfidz)/tahfidz-proker.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+  import ProkerBidangContainer from "~/features/proker/ProkerBidangContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("Tahfidz / Proker");
+</script>
+
+<template>
+  <Title>Tahfidz | Proker</Title>
+  <main>
+    <UCard>
+      <ProkerBidangContainer bidang="tahfidz" />
+    </UCard>
+  </main>
+</template>
