+++ app/pages/dashboard/(Tenaga Pendidik)/pendidik-proker.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+  import ProkerBidangContainer from "~/features/proker/ProkerBidangContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("Tenaga Pendidik / Proker");
+</script>
+
+<template>
+  <Title>Tenaga Pendidik | Proker</Title>
+  <main>
+    <UCard>
+      <ProkerBidangContainer bidang="tenaga_pendidik" />
+    </UCard>
+  </main>
+</template>
