+++ server/modules/absensi-generus/absensi-generus.dto.ts
@@ -4,9 +4,9 @@ import { kelasGenerusEnum } from "~~/shared/enum";
 
 export const OGenerusAbsensiForDesaList = z.object({
   ...OPagination.def.shape,
-  search: z.string(),
-  kelompokId: z.optional(z.coerce.number()),
-  kelasPengajian: z.enum(kelasGenerusEnum),
+  search: z.optional(z.string()),
+  kelompokId: z.optional(z.optional(z.coerce.number())),
+  kelasPengajian: z.optional(z.enum(kelasGenerusEnum)),
 });
 
 export type TGenerusAbsensiForDesaList = z.infer<
@@ -29,7 +29,7 @@ export const OGenerusAbsensiForDaerahList = z.object({
 
 export const OGenerusAbsensiForMudamudiList = z.object({
   ...OPagination.def.shape,
-  search: z.string(),
+  search: z.optional(z.string()),
   kelompokId: z.optional(z.coerce.number()),
   desaId: z.optional(z.coerce.number()),
 });
