+++ server/utils/dto/absensi.dto.ts
@@ -24,7 +24,7 @@ export type TAbsensiGenerusCreate = z.infer<
 export const OGenerusAbsensiList = z.object({
   ...OPagination.def.shape,
   search: z.optional(z.string()),
-  kelasPengajian: z.enum(kelasGenerusEnum),
+  kelasPengajian: z.optional(z.enum(kelasGenerusEnum)),
 });
 export type TGenerusAbsensiList = z.infer<typeof OGenerusAbsensiList>;
